🩹 Refactor: UsedTo 객체 코드를 다른 파일로 분리
